+++ topo_tools/core/assign/__init__.py
@@ -2,11 +2,12 @@
 
 from ._inputs import load_children, load_parent
 from ._many import assign_many
-from ._one import assign_one, prepare_parent_tiles
+from ._one import assign_one, child_bbox_extent, prepare_parent_tiles
 
 __all__ = [
     "assign_many",
     "assign_one",
+    "child_bbox_extent",
     "load_children",
     "load_parent",
     "prepare_parent_tiles",